moved App class to its own file, added documentation and linted.
